+++ CHANGELOG.md
@@ -7,6 +7,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ### Added
 - `normalizer --version` now specify if current version provide extra speedup (meaning mypyc compilation whl)
 
+### Removed
+- Breaking: Method `first()` and `best()` from CharsetMatch
+
 ## [3.0.0b1](https://github.com/Ousret/charset_normalizer/compare/2.1.0...3.0.0b1) (2022-08-15)
 
 ### Changed